feat: add --self-check for runtime dir and HTTP verification
Probe library/data read-write, open SQLite, serve /api/health, then exit. Also accepts self-check/doctor subcommands, and fixes linux/arm Statfs Bsize typing.

+++ internal/system/stats_linux.go
@@ -94,8 +94,8 @@ func diskUsage(path string) (DiskUsage, bool) {
 	if err := syscall.Statfs(path, &st); err != nil {
 		return DiskUsage{}, false
 	}
-	total := statBlocksBytes(st.Blocks, st.Bsize)
-	avail := statBlocksBytes(st.Bavail, st.Bsize)
+	total := statBlocksBytes(st.Blocks, int64(st.Bsize))
+	avail := statBlocksBytes(st.Bavail, int64(st.Bsize))
 	used := total - avail
 	pct := 0.0
 	if total > 0 {
